Team_2/Stress_Detector/requirements.txt
tensorflow-cpu==2.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
umap-learn>=0.5.0
PyYAML>=6.0
